fix mistakes about Dto's

+++ visa-tracking-service/src/main/java/com/visams/visatrackingservice/Dto/TravelLogDto.java
@@ -1,6 +1,5 @@
 package com.visams.visatrackingservice.Dto;
 
-import jakarta.persistence.Column;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -13,7 +12,7 @@ import java.time.LocalDate;
 public class TravelLogDto {
 
     private Integer logId;
-    private Integer touristId;
+    private Long touristId;
     private String location;
     private LocalDate checkInDate;
     private LocalDate checkOutDate;
