+++ visa-tracking-service/src/main/java/com/visams/visatrackingservice/Repository/TravelLogRepository.java
@@ -8,6 +8,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 public interface TravelLogRepository extends JpaRepository<TravelLog,Integer> {
 
     Page<TravelLog> findByLogId(Integer logId, Pageable pageable);
-    Page<TravelLog> findByTourist_touristId(Integer touristId, Pageable pageable);
+    Page<TravelLog> findByTourist_touristId(Long touristId, Pageable pageable);
     Page<TravelLog> findByLocationContainingIgnoreCase(String location, Pageable pageable);
 }
